Verify signatures when creating containers
On (docker create) / (docker run), try to use the user-provided
image specifier as a normalized reference or an image ID = config digest,
and verify the image with that identity and the maniest/signatures
stored in a previous commit.

If that verifies, things are obviously fine.  Do note that the image
specifier can be other things: a hostname-unqualified reference
found on one of the non-docker.io search registries, a unique prefix
of an image ID, name + image ID prefix.  If such a specifier is used,
it will either be refused as an invalid named reference, or just
not match the available signatures.

FIXME: refused names should not just abort!

For completely unnamed images, we invent an image ID = sha256(zero bytes).
A zero-byte config.json is invalid, so this should not match any existing
image.

Ultimately, though, docker-daemon: collapses all ID references into the
root "" policy configuration identity: the user can either allow all
of them, or reject them all, as a class.  Presumably, anyone
setting up policy enforcement would reject all ID references.

NOTE: The (docker build) only uses image IDs, and thus uses only
the default "" policy identity (both for the initial FROM $something image,
and for successive layers).  This means, that if the default policy is
tightened to require signatures, (docker build) is prevented.  This seems
appropriate for production configurations.

Note that the enforcement happens when _creating_ the container; not
starting a created one.  In most situations this should not make
a difference.  Ultimately, if the worry is that a previously acceptable
container is no longer acceptable, checking on creation/start time is
equally insufficient; a regular audit af running containers is be necessary.

This should not change behavior by default, because the default
policy we ship specifically configures docker-daemon: to be insecureAcceptAnything.

WARNING: This DOES NOT BUILD because it references sirupsen/logrus, not
Sirupsen/logrus; and docker/docker/client.